src/requirements.txt
pillow==10.4.0
boto3==1.35.0
ruff==0.6.1